Move the consent BFF tests to access-management-bff
These tests import only from clients/access-management-bff/consent and
building-blocks/access-management-bff/consent, nothing from authentication,
so tests/authentication/consent/bff put them under the wrong domain. Both
the client and the building block trees already have an access-management-bff
folder at the top level, and the test tree mirrors those trees elsewhere.

Relative imports lose one level, since the new path is one shallower. base_dir
in smoke.yaml follows the move.

Left namespace as authentication. It becomes the Kubernetes namespace of the
generated CronJob, so changing it needs the namespace to exist cluster side
first.
