+++ K6/api/tests/access-management-bff/consent/consent-requests-worst-case.js
@@ -1,6 +1,6 @@
-import { getItemFromList, getOptions } from "../../../../../helpers.js";
-import { GetConsentLog } from "../../../../building-blocks/access-management-bff/consent/index.js";
-import { ConsentDomainChecks } from "../../../../domain-checks/access-management/consent.js";
+import { getItemFromList, getOptions } from "../../../../helpers.js";
+import { GetActiveConsents } from "../../../building-blocks/access-management-bff/consent/index.js";
+import { ConsentDomainChecks } from "../../../domain-checks/access-management/consent.js";
 import { getClients, getTokenOpts, worst_case_users as users } from "./commons.js";
 
 // One threshold per user, so the summary breaks the numbers out per user rather
@@ -8,7 +8,8 @@ import { getClients, getTokenOpts, worst_case_users as users } from "./commons.j
 export const options = getOptions(users.map(user => { return { unique_id: user.label }; }));
 
 /**
- * Test: reading the consent log as the users that have the most consent requests.
+ * Test: reading the active consents as the users that have the most consent
+ * requests.
  *
  * The users are walked in order rather than drawn at random, so every one of them
  * is read the same number of times and their numbers stay comparable.
@@ -20,10 +21,10 @@ export default function () {
 
     tokenGenerator.setTokenGeneratorOptions(getTokenOpts(from.userId, from.partyUuid));
 
-    const log = GetConsentLog(consentClient, from.partyUuid, { unique_id: from.label });
+    const activeConsents = GetActiveConsents(consentClient, from.partyUuid, { unique_id: from.label });
 
-    ConsentDomainChecks.CheckConsentResponse(log, "GetConsentLog");
+    ConsentDomainChecks.CheckConsentResponse(activeConsents, "GetActiveConsents");
 }
 
 // add the custom reporting for this test to the default summary
-export { handleSummary } from "../../../../../common-imports.js";
+export { handleSummary } from "../../../../common-imports.js";
